Cue-Fit_Service_FrontEnd/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
asyncer==0.0.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
gradio==4.7.1
gradio_client==0.7.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.0.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyMatting==1.1.11
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
rembg==2.0.52
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tifffile==2023.9.26
tomlkit==0.12.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
watchdog==3.0.0
wcwidth==0.2.12
websockets==11.0.3
xlrd==2.0.1
yarl==1.9.3
